updating and working on relationships
so you have proper nouns, that are the links mostly, and improper nouns - these have to be delicately split and divided, to define the relationships. what i am doing is breaking a fragment of the node graph off and defining the relationship it has...then putting it back together in which it forms from the paragraph.