Data is cleaned
I have one more process to go and that is relationships- I saved it to an xml file for further processing with ChatGPT open AI to handle the last four identities and that is child, parent, relationship and node graph. There was just too many lines of code to keep going in one notebook.